feat: Implement IPAM subnets and refactor device form
Introduces a new `Subnet` data type and database table to support IP address management (IPAM).

This change also refactors the `DeviceForm` to accept `Partial<NetworkDevice>` for `initialData`. This allows for pre-filling device details, such as just the IP address, when creating a new device linked to a customer. The `isEditMode` logic has been updated to correctly reflect whether an existing device is being edited or a new one is being created.